뷰 추가 및 CommonsController 생성
Oracle_Yomul.sql
- 댓글 정보 뷰에 이미지 추가
- faq, qna 뷰 생성

CommonsController.java
- 공용으로 쓰이는 컨트롤러 생성
- 댓글 작성/좋아요/신고/댓글 페이지 이동

NearController.java
- CommonsController로 옮긴 함수 삭제

near_info.jsp
- 댓글 작성 ajax 추가(미완)
