+++ Yomul/src/main/java/com/yomul/yomul/NearController.java
@@ -143,77 +143,6 @@ public class NearController {
 		return mv;
 	}
 	
-	// 내 근처 상세페이지 댓글 작성 ajax
-	@ResponseBody
-	@RequestMapping(value = "/near_info/write_comment", method = RequestMethod.GET)
-	public String write_comment(String ano, String content, HttpSession session) {
-		// 로그인 정보 불러오기
-		MemberVO member = (MemberVO) session.getAttribute("member");
-		String mno = member.getNo();
-		
-		// 댓글 정보 세팅
-		CommentVO vo = new CommentVO();
-		vo.setArticleNo(ano);
-		vo.setWriter(mno);
-		vo.setContent(content);
-		
-		// DB에 댓글 저장
-		int result = commentService.addComment(vo);
-		
-		// 성공 시 1, 실패 시 0 반환
-		return String.valueOf(result);
-	}
-	
-	// 내 근처 상세페이지 댓글 좋아요 ajax
-	@ResponseBody
-	@RequestMapping(value = "/near_info/insert_like", method = RequestMethod.GET)
-	public String insert_like(String ano, HttpSession session) {
-		// 로그인 정보 불러오기
-		MemberVO member = (MemberVO) session.getAttribute("member");
-		String mno = member.getNo();
-		
-		if(mno.equals("")) { // 로그인 안한 경우 -1 반환
-			return "-1";
-		}
-		
-		// DB에 좋아요 데이터 저장
-		int result = likeService.insertLike(ano, mno);
-		
-		// 성공 시 1, 실패 시 0 반환
-		return String.valueOf(result);
-	}
-	
-	// 내 근처 상세페이지 댓글 신고 추가 ajax
-	@ResponseBody
-	@RequestMapping(value = "/near_info/insert_report", method = RequestMethod.GET)
-	public String insert_report(String ano, HttpSession session) {
-		// 로그인 정보 불러오기
-		MemberVO member = (MemberVO) session.getAttribute("member");
-		String mno = member.getNo();
-
-		if(mno.equals("")) { // 로그인 안한 경우
-			return "-1";
-		}
-		
-		return String.valueOf(reportService.insertReport(ano, mno));
-	}
-	
-	// 댓글 페이지 이동 ajax
-	@ResponseBody
-	@RequestMapping(value = "/near_info/comments", method = RequestMethod.GET, produces = "text/plain; charset=utf8")
-	public String near_info_comment(int no, int page) {
-		int commentCount = commentService.getCommentCount("N" + no);
-		ArrayList<CommentVO> comments = commentService.getCommentList("N" + no, page);
-		HashMap<String, Integer> commentPageInfo = Commons.getPageInfo(page, commentCount, 10);
-		
-		// json으로 변환
-		String commentsJson = Commons.parseJson(comments);
-		String pageInfoJson = Commons.parseJson(commentPageInfo);
-		
-		// 두 json 객체를 배열에 담아 반환
-		return "[" + commentsJson + "," + pageInfoJson + "]";
-	}
-	
 	// 아직 미완 로그인 기능 구현되면 구현한 예정
 	// 단골 등록 ajax
 	@ResponseBody
